Update replication checkpoint documents to version 3.
Update replication checkpoint documents to version 3 of the
replication protocol. Previous replication documents will
be transformed into the new version before being saved. Transforming
previous documents into the new format will result in a single history
entry for **all** previous replications, this is because we only have
the previous last update seq in the local documents.

Documents are updated on save rather than through a migration is to keep
both the local and remote documents using the same format. If the
`source_last_seq` is not present in the document, it is assumed that
the document is in the old format `{"seq":500}` rather than the new format.